+++ triplea/service/repository/state/custom/extract_kg_abstract.py
@@ -4,14 +4,14 @@ from triplea.client.triple_extraction import extract_triple
 import triplea.service.repository.persist as persist
 
 
-def extract_triple_abstract_save(article: Article):
+def extract_triple_abstract_save(article: Article,article_id):
     article.FlagExtractKG = 1
     if article.Abstract is not None:
         # triples_list = extract_triples(article.Abstract)  # Expire Module
+        triples_list = []
         triples_list = extract_triple(article.Abstract)
-        triples_list = []  # CRITICAL must be API
         for t in triples_list:
-            t["PMID"] = article.PMID
+            t["Article_ID"] = article_id
             persist.create_triple(t)
 
     return article