+++ triplea/the_private_backyard2.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 # noqa: F401
 
+from bson import BSON, ObjectId
 import click
 import time
 import sys
@@ -32,38 +33,54 @@ from triplea.service.repository.pipeline_flag import (
 from triplea.service.repository.state.custom.affiliation_mining import (
     get_affiliation_structured,
 )
-
+from triplea.service.repository.state.initial_arxiv import get_article_list_from_arxiv_all_store_to_arepo
+from triplea.service.repository.state.initial import get_article_list_from_pubmed_all_store_to_arepo
 if __name__ == "__main__":
     pass
-    export_triplea_csvs_in_relational_mode_save_file(
-        "export.csv", proccess_bar=True, limit_sample=0
-    )
 
+    # Pipeline Sample
+
+    # # Step 1 - Get article from Arxiv
+    # arxiv_search_string = 'ti:"large language model" AND ti:Benchmark'
+    # get_article_list_from_arxiv_all_store_to_arepo(arxiv_search_string,0,10)
+
+    # # Step 2 - Get article from Pubmed
+    # pubmed_search_string = '("large language model"[Title]) AND (Benchmark[Title/Abstract])'
+    # get_article_list_from_pubmed_all_store_to_arepo(pubmed_search_string)
+ 
+
+
+    # Step 3 - Get info
+    persist.print_article_info_from_repo()
+
+    # # Step 4 - Moving from `0` to `1`
+    # move_state_forward(0)                    
+
+    # # Step 5 - Moving from `1` to `2`
+    # move_state_forward(1)
+                    
+    # # Step 6 - Moving from `2` to `3`
     # move_state_forward(2)
-    # go_affiliation_mining()
+                    
+    # Get article info
+    # print()               
+    # persist.print_article_short_description("37567487","pmid")
 
-    # persist.change_flag_extract_topic(1,0)
-    # go_extract_topic(proccess_bar=True)
-
-    # aff_text = "Institute for Molecular Medicine Finland (FIMM), Helsinki Institute of Life Science (HiLIFE), University of Helsinki, Helsinki, Finland. aarno.palotie@helsinki.fi"
-    # aff_text = "Department of Neurology and Institute of Neurology, Huashan Hospital, State Key Laboratory of Medical Neurobiology and MOE Frontiers Center for Brain Science, Shanghai Medical College, Fudan University, National Center for Neurological Disorders, Shanghai, China. jintai_yu@fudan.edu.cn"
-    # aff_text = "Department of Ophthalmology, University of Washington, Seattle, Washington, USA"
-    # print(get_affiliation_structured(aff_text))
-
-    # from triplea.service.repository.state.custom.affiliation_mining import _is_country
-    # print(_is_country("Finland. aarno.palotie@helsinki.fi"))
-    # print(_is_country("Finland"))
-
-    # print(parse_affiliation(aff_text))
-
-    # go_affiliation_mining(method='Titipata')
-
-    # import triplea.service.repository.state as state_manager
-    # a = persist.get_article_by_pmid('31679581')
+    
+    # a = persist.get_article_by_id(ObjectId('658f85228f23534d63358a19'))
     # updated_article = Article(**a.copy())
-    # state_manager.affiliation_mining_titipata(updated_article)
+    # print(updated_article.State)
+    # print(updated_article.Title)
+    # print(updated_article.Published)
+    # print(type(updated_article.Published))
 
-    # import triplea.service.repository.state as state_manager
-    # a = persist.get_article_by_pmid('34358588')
-    # updated_article = Article(**a.copy())
-    # state_manager.parsing_details(updated_article)
+    import triplea.service.repository.pipeline_flag as cPIPELINE
+
+    # cPIPELINE.go_extract_topic()
+
+    # cPIPELINE.go_affiliation_mining(method="Titipata")
+
+    # cPIPELINE.go_extract_triple()
+
+    export_triplea_csvs_in_relational_mode_save_file("export")
+